unused/barged_api - Copy/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth-oauthlib
google-generativeai
